Fixed Windows Errors For Current Build
INCLUDE <memory> for god sakes.
Also fixed ascpect ratio.

+++ app/ShatteredMoon_25/src/main.cpp
@@ -16,7 +16,7 @@ int main(int argc, char **argv) {
 	glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
 	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
 	glfwWindowHint(GLFW_RESIZABLE, GL_FALSE);
-	int windowWidth = 1600, windowHeight = 900;
+	float windowWidth = 1600.0f, windowHeight = 900.0f;
 	GLFWwindow * window = glfwCreateWindow(windowWidth, windowHeight, "LearnOpenGL", nullptr, nullptr);
 	if (window == nullptr)
 	{
